Refactor load config saga
-  Move code to the respective staging areas

JIRA: RAIL-3383
