+++ libs/sdk-ui-dashboard/src/model/commandHandlers/loadDashboard/loadDashboardConfig.ts
@@ -1,25 +1,16 @@
 // (C) 2021 GoodData Corporation
-import includes from "lodash/includes";
 import { DashboardConfig, DashboardContext } from "../../types/commonTypes";
-import {
-    IDateFilterConfig,
-    IDateFilterConfigsQueryResult,
-    ISettings,
-    IUserWorkspaceSettings,
-} from "@gooddata/sdk-backend-spi";
+import { IDateFilterConfigsQueryResult, IUserWorkspaceSettings } from "@gooddata/sdk-backend-spi";
 import { LoadDashboard } from "../../commands/dashboard";
 import { all, call } from "redux-saga/effects";
 import { dateFilterValidationFailed } from "../../events/dashboard";
 import { defaultDateFilterConfig } from "../../_staging/dateFilterConfig/defaultConfig";
-import {
-    DateFilterConfigValidationResult,
-    filterOutWeeks,
-    validateDateFilterConfig,
-} from "../../_staging/dateFilterConfig/validation";
 import { eventDispatcher } from "../../eventEmitter/eventDispatcher";
 import { IColorPalette } from "@gooddata/sdk-model";
 import { PromiseFnReturnType } from "../../types/sagas";
 import { ILocale } from "@gooddata/sdk-ui";
+import { getValidDateFilterConfig } from "../../_staging/dateFilterConfig/validation";
+import { stripUserAndWorkspaceProps } from "../../_staging/settings/conversion";
 
 function loadDateFilterConfig(ctx: DashboardContext): Promise<IDateFilterConfigsQueryResult | undefined> {
     const { backend, workspace } = ctx;
@@ -45,36 +36,6 @@ function loadColorPalette(ctx: DashboardContext): Promise<IColorPalette> {
     const { backend, workspace } = ctx;
 
     return backend.workspace(workspace).styling().getColorPalette();
-}
-
-const FallbackToDefault: DateFilterConfigValidationResult[] = ["ConflictingIdentifiers", "NoVisibleOptions"];
-
-/**
- * Given the date filter config loaded from backend and the settings, this function will perform validation
- * of the config and if needed also cleanup of invalid/disabled presets.
- */
-function getValidDateFilterConfig(
-    config: IDateFilterConfig,
-    settings: ISettings,
-): [IDateFilterConfig, DateFilterConfigValidationResult] {
-    const configValidation = validateDateFilterConfig(config);
-    const validConfig = !includes(FallbackToDefault, configValidation) ? config : defaultDateFilterConfig;
-
-    const dateFilterConfig = !settings.enableWeekFilters ? filterOutWeeks(validConfig) : validConfig;
-
-    return [dateFilterConfig, configValidation];
-}
-
-function stripUserAndWorkspaceProps(userWorkspaceSettings: IUserWorkspaceSettings): ISettings {
-    const {
-        userId: _userId,
-        locale: _locale,
-        separators: _separators,
-        workspace: _workspace,
-        ...rest
-    } = userWorkspaceSettings;
-
-    return rest;
 }
 
 /**
